move benchmarks to a separate file
